requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
contourpy==1.3.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.18.0
Flask==2.3.3
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==20.1.0
huggingface-hub==0.30.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.7.1
mlflow==2.5.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.12.0.88
packaging==23.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pyarrow==12.0.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.3
scikit-learn==1.2.0
scipy==1.13.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.64.1
transformers==4.28.0
triton==3.1.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
zipp==3.21.0
